+++ src/main/java/com/MaxBr221/GitHub/dtos/entitysDTO/ServicoRequestDTO.java
@@ -2,9 +2,11 @@ package com.MaxBr221.GitHub.dtos.entitysDTO;
 
 import com.MaxBr221.GitHub.model.Status;
 
+import java.math.BigDecimal;
+
 public record ServicoRequestDTO(
         String nome,
-        double preco,
+        BigDecimal preco,
         Status status,
         String descricao
 ) {
